Moved stuff to mol-util/array.ts
