+++ src/extensions/mvs/load-helpers.ts
@@ -10,6 +10,7 @@ import { StructureFromModel, TransformStructureConformation } from '../../mol-pl
 import { StructureRepresentation3D } from '../../mol-plugin-state/transforms/representation';
 import { PluginContext } from '../../mol-plugin/context';
 import { StateBuilder, StateObjectSelector, StateTransformer } from '../../mol-state';
+import { arrayDistinct } from '../../mol-util/array';
 import { MVSAnnotationColorThemeProps, MVSAnnotationColorThemeProvider } from './additions/annotation-color-theme';
 import { MVSAnnotationLabelRepresentationProvider } from './additions/annotation-label/representation';
 import { MVSAnnotationSpec } from './additions/annotation-prop';
@@ -19,7 +20,7 @@ import { CustomTooltipsProps } from './additions/custom-tooltips-prop';
 import { MultilayerColorThemeName, MultilayerColorThemeProps, NoColor } from './additions/multilayer-color-theme';
 import { SelectorAll } from './additions/selector';
 import { rowToExpression, rowsToExpression } from './helpers/selections';
-import { ElementOfSet, canonicalJsonString, decodeColor, distinct, isDefined, stringHash } from './helpers/utils';
+import { ElementOfSet, canonicalJsonString, decodeColor, isDefined, stringHash } from './helpers/utils';
 import { MolstarLoadingContext } from './load';
 import { Kind, ParamsOfKind, SubTree, SubTreeOfKind, Tree, getChildren } from './tree/generic/tree-schema';
 import { dfs } from './tree/generic/tree-utils';
@@ -136,7 +137,7 @@ export function collectAnnotationTooltips(tree: SubTreeOfKind<MolstarTree, 'stru
             };
         }
     });
-    return distinct(annotationTooltips);
+    return arrayDistinct(annotationTooltips);
 }
 /** Collect annotation tooltips from all nodes in `tree`. */
 export function collectInlineTooltips(tree: SubTreeOfKind<MolstarTree, 'structure'>, context: MolstarLoadingContext) {
