+++ src/mol-util/array.ts
@@ -1,9 +1,11 @@
 /**
- * Copyright (c) 2018-2021 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author Adam Midlik <midlik@gmail.com>
  */
 
+import { Json, canonicalJsonString } from '../extensions/mvs/helpers/utils';
 import { NumberArray } from './type-helpers';
 
 // TODO move to mol-math as Vector???
@@ -155,4 +157,74 @@ export function arrayMapUpsert<T>(xs: [string, T][], key: string, value: T) {
         }
     }
     xs.push([key, value]);
+}
+
+/** Return an array containing integers from [start, end) if `end` is given,
+ * or from [0, start) if `end` is omitted. */
+export function range(start: number, end?: number): number[] {
+    if (end === undefined) {
+        end = start;
+        start = 0;
+    }
+    const length = Math.max(end - start, 0);
+    const result = Array(length);
+    for (let i = 0; i < length; i++) {
+        result[i] = start + i;
+    }
+    return result;
+}
+
+/** Copy all elements from `src` to the end of `dst`.
+ * Equivalent to `dst.push(...src)`, but avoids storing element on call stack. Faster that `extend` from Underscore.js.
+ * `extend(a, a)` will double the array.
+ */
+export function arrayExtend<T>(dst: T[], src: ArrayLike<T>): void {
+    const offset = dst.length;
+    const nCopy = src.length;
+    dst.length += nCopy;
+    for (let i = 0; i < nCopy; i++) {
+        dst[offset + i] = src[i];
+    }
+}
+
+/** Check whether `array` is sorted, sort if not. */
+export function sortIfNeeded<T>(array: T[], compareFn: (a: T, b: T) => number): T[] {
+    for (let i = 1, n = array.length; i < n; i++) {
+        if (compareFn(array[i - 1], array[i]) > 0) {
+            return array.sort(compareFn);
+        }
+    }
+    return array;
+}
+
+/** Remove all elements from the array which do not fulfil `predicate`. Return the modified array itself. */
+export function filterInPlace<T>(array: T[], predicate: (x: T) => boolean): T[] {
+    const n = array.length;
+    let iDest = 0;
+    for (let iSrc = 0; iSrc < n; iSrc++) {
+        if (predicate(array[iSrc])) {
+            array[iDest++] = array[iSrc];
+        }
+    }
+    array.length = iDest;
+    return array;
+}
+
+/** Return an array of all distinct values from `values`
+ * (i.e. with removed duplicates).
+ * Uses deep equality for objects and arrays,
+ * independent from object key order and undefined properties.
+ * E.g. {a: 1, b: undefined, c: {d: [], e: null}} is equal to {c: {e: null, d: []}}, a: 1}.
+ * If two or more objects in `values` are equal, only the first of them will be in the result. */
+export function arrayDistinct<T extends Json>(values: T[]): T[] {
+    const seen = new Set<string>();
+    const result: T[] = [];
+    for (const value of values) {
+        const key = canonicalJsonString(value);
+        if (!seen.has(key)) {
+            seen.add(key);
+            result.push(value);
+        }
+    }
+    return result;
 }
